Feature/dtype achsen und precision (#14)
* add test_verify.py

verbesserung verify

fp16→fp16: max 0.222, rel_L2 8.4e-4
fp8e5m2→fp16: max 0.163, rel_L2 6.1e-4

* feature pipeline kann fp16 jetzt rechnen

* feature tf32

(input 32-Bit-fp32 -> müssen mitten drin in tfloat32 umgewandelt werden für tensor-cores).

sonst richtig rechnen aber langsamer

* feature fp8

input wird als fp16 gebaut und dann auf fp8 runtergecastet

* add Zahlenformate in UI

* add charts

* features:

- render status jetzt ganz oben
- beide diagramme jetzt übereinander statt nebeneinander
- tabs für variaten

* bug + visual fixes

* add Tz4

+++ project/results/kernels/ik_kj_to_ij__bf16-fp32__TM128_TN128_TK64.py
@@ -2,12 +2,12 @@
 # Auto-generiert vom cuTile Performance Lab (Codegen C1).
 # Aus einer RunConfig erzeugt.
 # Ausdruck : ik,kj->ij
-# Format   : fp16 -> fp32 (Akku)
+# Format   : bf16 -> fp32 (Akku)
 # Tile     : TM=128 TN=128 TK=64 | swizzle=False
 # ==========================================================================
 """Generierter cuTile-GEMM (Codegen C1) — Kontraktion ik,kj->ij.
 
-Input-dtype: fp16 (Laufzeit-torch-dtype, steht NICHT im Kernel-Koerper).
+Input-dtype: bf16 (Laufzeit-torch-dtype, steht NICHT im Kernel-Koerper).
 Akkumulator: fp32 (ct.float32).
 Tile-Literale: TM=128, TN=128, TK=64 (fest in den Quelltext gebacken).
